--- a/cpp/src/arrow/util/decimal_benchmark.cc
+++ b/cpp/src/arrow/util/decimal_benchmark.cc
@@ -26,19 +26,51 @@
 namespace arrow {
 namespace Decimal {
 
-static void FromString(benchmark::State& state) {  // NOLINT non-const reference
-  std::vector<std::string> values = {"0",
-                                     "1.23",
-                                     "12.345e6",
-                                     "-12.345e-6",
-                                     "123456789.123456789",
-                                     "1231234567890.451234567890"};
+static const std::vector<std::string>& GetValuesAsString() {
+  static const std::vector<std::string>* values =
+      new std::vector<std::string>{"0",
+                                   "1.23",
+                                   "12.345e6",
+                                   "-12.345e-6",
+                                   "123456789.123456789",
+                                   "1231234567890.451234567890"};
+  return *values;
+}
 
+static std::vector<std::pair<Decimal128, int32_t>> BuildDecimalValuesAndScales() {
+  const std::vector<std::string>& value_strs = GetValuesAsString();
+  std::vector<std::pair<Decimal128, int32_t>> result(value_strs.size());
+  for (size_t i = 0; i < value_strs.size(); ++i) {
+    int32_t precision;
+    Decimal128::FromString(value_strs[i], &result[i].first, &result[i].second,
+                           &precision);
+  }
+  return result;
+}
+
+static const std::vector<std::pair<Decimal128, int32_t>>& GetDecimalValuesAndScales() {
+  static const auto* values =
+      new std::vector<std::pair<Decimal128, int32_t>>(BuildDecimalValuesAndScales());
+  return *values;
+}
+
+static void FromString(benchmark::State& state) {  // NOLINT non-const reference
+  const std::vector<std::string>& values = GetValuesAsString();
   for (auto _ : state) {
     for (const auto& value : values) {
       Decimal128 dec;
       int32_t scale, precision;
       benchmark::DoNotOptimize(Decimal128::FromString(value, &dec, &scale, &precision));
+    }
+  }
+  state.SetItemsProcessed(state.iterations() * values.size());
+}
+
+static void ToString(benchmark::State& state) {  // NOLINT non-const reference
+  const auto& values = GetDecimalValuesAndScales();
+  for (auto _ : state) {
+    for (const auto& item : values) {
+      benchmark::DoNotOptimize(item.first.ToString(item.second));
     }
   }
   state.SetItemsProcessed(state.iterations() * values.size());
@@ -158,6 +190,7 @@ static void BinaryBitOp(benchmark::State& state) {  // NOLINT non-const referenc
 }
 
 BENCHMARK(FromString);
+BENCHMARK(ToString);
 BENCHMARK(BinaryMathOp);
 BENCHMARK(BinaryMathOpAggregate);
 BENCHMARK(BinaryCompareOp);