Added a benchmark to decimal_benchmark.

```
Running release/arrow-decimal-benchmark
Run on (12 X 4500 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x6)
  L1 Instruction 32 KiB (x6)
  L2 Unified 1024 KiB (x6)
  L3 Unified 8448 KiB (x1)
Load Average: 0.23, 0.22, 0.23
----------------------------------------------------------------------------------
Benchmark                        Time             CPU   Iterations UserCounters...
----------------------------------------------------------------------------------
ToString (before)             2205 ns         2205 ns       321603 items_per_second=2.72094M/s
ToString (after)               216 ns          216 ns      3235082 items_per_second=27.7865M/s
```